Rename files to have d2 prefix
