+++ SlashGaming-Diablo-II-API/include/c/game_struct.h
@@ -38,6 +38,6 @@
 #ifndef SGD2MAPI_C_GAME_STRUCT_H_
 #define SGD2MAPI_C_GAME_STRUCT_H_
 
-#include "game_struct/unicode_char.h"
+#include "game_struct/d2_unicode_char.h"
 
 #endif // SGD2MAPI_C_GAME_STRUCT_H_
