+++ SlashGaming-Diablo-II-API/src/cxx/game_struct/d2_unicode_char_impl.hpp
@@ -35,8 +35,8 @@
  *  work.
  */
 
-#ifndef SGD2MAPI_CXX_GAME_STRUCT_UNICODE_CHAR_IMPL_HPP_
-#define SGD2MAPI_CXX_GAME_STRUCT_UNICODE_CHAR_IMPL_HPP_
+#ifndef SGD2MAPI_CXX_GAME_STRUCT_D2_UNICODE_CHAR_IMPL_HPP_
+#define SGD2MAPI_CXX_GAME_STRUCT_D2_UNICODE_CHAR_IMPL_HPP_
 
 #include <cstdint>
 
@@ -52,4 +52,4 @@ namespace d2 {
 
 } // namespace d2
 
-#endif // SGD2MAPI_CXX_GAME_STRUCT_UNICODE_CHAR_IMPL_HPP_
+#endif // SGD2MAPI_CXX_GAME_STRUCT_D2_UNICODE_CHAR_IMPL_HPP_
